1. 修正esp32 pwm接口处理

+++ hal/src/atom/pwm_hal.c
@@ -43,12 +43,12 @@ static pwm_state_t PWM_State[TIM_NUM] = {
 
 void HAL_PWM_Write(uint16_t pin, uint8_t value)
 {
-    HAL_PWM_Write_With_Frequency_Ext(pin, (uint16_t)value, configTIM_PWM_FREQ);
+    HAL_PWM_Write_With_Frequency_Ext(pin, (uint32_t)value, configTIM_PWM_FREQ);
 }
 
 void HAL_PWM_Write_With_Frequency(uint16_t pin, uint8_t value, uint16_t pwm_frequency)
 {
-    HAL_PWM_Write_With_Frequency_Ext(pin, (uint16_t)value, pwm_frequency);
+    HAL_PWM_Write_With_Frequency_Ext(pin, (uint32_t)value, pwm_frequency);
 }
 
 void HAL_PWM_Write_Ext(uint16_t pin, uint32_t value)
